requirements.txt
numpy==1.25.0
pandas==1.5.3
pymysql==1.0.3
python_dateutil==2.8.2
pytz==2021.1
requests==2.28.1
todoist_api_python==2.0.1
loguru==0.6.0
pytube==12.1.0
msal==1.20.0
python-dotenv==0.21.1
quarter_lib==0.0.25
PyGithub==1.58.2
google-api-python-client==2.88.0
google-auth-oauthlib==1.0.0
python-telegram-bot==20.0
uvicorn==0.23.2
fastapi==0.103.0